Paczka Poprawek
Nowszy silnik + Mapa

+++ svn — reward box/forgottenserver-master/src/database.cpp
@@ -1,6 +1,6 @@
 /**
  * The Forgotten Server - a free and open-source MMORPG server emulator
- * Copyright (C) 2015  Mark Samman <mark.samman@gmail.com>
+ * Copyright (C) 2016  Mark Samman <mark.samman@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -191,7 +191,7 @@ DBResult::DBResult(MYSQL_RES* res)
 {
 	handle = res;
 
-	int32_t i = 0;
+	size_t i = 0;
 
 	MYSQL_FIELD* field = mysql_fetch_field(handle);
 	while (field) {
