+++ svn — reward box/forgottenserver-master/src/database.h
@@ -1,6 +1,6 @@
 /**
  * The Forgotten Server - a free and open-source MMORPG server emulator
- * Copyright (C) 2015  Mark Samman <mark.samman@gmail.com>
+ * Copyright (C) 2016  Mark Samman <mark.samman@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -179,7 +179,7 @@ class DBResult
 		MYSQL_RES* handle;
 		MYSQL_ROW row;
 
-		std::map<std::string, uint32_t> listNames;
+		std::map<std::string, size_t> listNames;
 
 	friend class Database;
 };
